Ensure we return an integer
PDO and Doctrine return a string, but we pledge to return an integer.

+++ src/Datagrid/Pager.php
@@ -45,7 +45,7 @@ class Pager extends BasePager
             current($this->getCountColumn())
         ));
 
-        return $countQuery->resetDQLPart('orderBy')->getQuery()->getSingleScalarResult();
+        return (int) ($countQuery->resetDQLPart('orderBy')->getQuery()->getSingleScalarResult());
     }
 
     public function getResults($hydrationMode = Query::HYDRATE_OBJECT)
